- Added one more screenshot test and clean up Playground activity
