+++ app/src/androidTest/java/com/kickstarter/screenshoot/testing/ui/components/TextViewElementShotTest.kt
@@ -111,4 +111,32 @@ class TextViewElementShotTest : ScreenshotTest {
         textView.text = element.getStyledComponents(body, headerSize, app)
         compareScreenshot(textView)
     }
+
+    @Test
+    fun textElementVeryComplexList() {
+        val textView = (
+            LayoutInflater.from(InstrumentationRegistry.getInstrumentation().targetContext).inflate(
+                R.layout.view_element_text_from_html, null
+            ) as ConstraintLayout
+            ).findViewById(R.id.text_view) as AppCompatTextView
+
+        val html = "<ul>\n" +
+            "   <li>This</li>\n" +
+            "   <li>More text here <em><strong>lalalalalala</strong></em> and here </li>\n" +
+            "   <li><strong>is</strong></li>\n" +
+            "   <li>is <strong>is</strong> is</li>\n" +
+            "   <li><em><strong>is</strong></em> with some <strong>text</strong> in the middle <em><strong>is</strong></em></li>\n" +
+            "   <li><em>a</em></li>\n" +
+            "   <li><a href=\\\"http://record.pt\\\" target=\\\"_blank\\\" rel=\\\"noopener\\\">link</a></li>\n" +
+            "   <li>text with <a href=\\\"http://record.pt\\\" target=\\\"_blank\\\" rel=\\\"noopener\\\">link</a></li>\n" +
+            "   <li>Hola <strong><em>que tal </em></strong> majete</li>\n" +
+            "</ul>"
+
+        val listOfElements = HTMLParser().parse(html)
+
+        val element = listOfElements.first() as TextViewElement
+
+        textView.text = element.getStyledComponents(body, headerSize, app)
+        compareScreenshot(textView)
+    }
 }
