+++ app/src/internal/java/com/kickstarter/ui/activities/PlaygroundActivity.kt
@@ -3,7 +3,6 @@ package com.kickstarter.ui.activities
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.text.method.LinkMovementMethod
 import android.util.Pair
 import android.view.View
 import androidx.annotation.RequiresApi
@@ -11,9 +10,6 @@ import com.kickstarter.R
 import com.kickstarter.databinding.PlaygroundLayoutBinding
 import com.kickstarter.libs.BaseActivity
 import com.kickstarter.libs.RefTag
-import com.kickstarter.libs.htmlparser.HTMLParser
-import com.kickstarter.libs.htmlparser.TextViewElement
-import com.kickstarter.libs.htmlparser.getStyledComponents
 import com.kickstarter.libs.qualifiers.RequiresActivityViewModel
 import com.kickstarter.mock.factories.ProjectFactory
 import com.kickstarter.models.Project
@@ -33,39 +29,6 @@ class PlaygroundActivity : BaseActivity<PlaygroundViewModel.ViewModel?>() {
         binding = PlaygroundLayoutBinding.inflate(layoutInflater)
         view = binding.root
         setContentView(view)
-
-        val context = this
-
-        // - Allow clickable spans
-        binding.text.linksClickable = true
-        binding.text.isClickable = true
-        binding.text.movementMethod = LinkMovementMethod.getInstance()
-
-        val headerSize = resources.getDimensionPixelSize(R.dimen.title_3)
-        val body = resources.getDimensionPixelSize(R.dimen.callout)
-        val html = "<ul>\n" +
-            "   <li>This</li>\n" +
-            "   <li><em><strong>lalalalalala</strong></em></li>\n" +
-            "   <li><strong>is</strong></li>\n" +
-            "   <li><em>a</em></li>\n" +
-            "   <li><a href=\\\"http://record.pt\\\" target=\\\"_blank\\\" rel=\\\"noopener\\\">list</a></li>\n" +
-            "   <li>Hola <strong> que tal </strong> majete</li>\n" +
-            "</ul>"
-
-        val html4 = "<ul>\n" +
-            "   <li> <em><strong>Epic</strong></em> <em><strong>Treasures</strong></em><strong> (\$6 PDF, \$13 Print/PDF bundle):</strong> Get 26 pages of marvelous magic items that become even more powerful in the hands of characters with divine blessings and epic boons! Don the <em>cuirass of miracles</em>, seek the <em>golden fleece</em>, sow the <em>teeth</em> <em>of the hydra, </em>and harness the <em>yoke of the brazen bull!</em> </li>\n" +
-            "   <li> <em><strong>Sea Monsters</strong></em> <strong>(\$15</strong> <strong>PDF, \$20 Print/PDF Bundle):</strong> Over 60 maritime monstrosities and nautical nemeses for your heroes, from low-level minions of the deep like mutant <strong>selachim sahuagin</strong> and to <strong>reef hags</strong>, to vast <strong>living islands </strong>and half-alive <strong>coral golems</strong>, plus savage predators like the <strong>devilfish </strong>and <strong>slaughtermaw lamprey</strong>, and legendary foes like the <strong>scylla, charybdis, </strong>and even the <strong>Midgard Serpent</strong>.</li>\n" +
-            "</ul>"
-
-        val listOfElements = HTMLParser().parse(html)
-        val element = listOfElements.first() as TextViewElement
-        binding.text.text = element.getStyledComponents(body, headerSize, this)
-
-        val listOfElements2 = HTMLParser().parse(html4)
-
-        val element2 = listOfElements2.first() as TextViewElement
-
-        binding.text2.text = element2.getStyledComponents(body, headerSize, this)
 
         setStepper()
         setProjectActivityButtonClicks()
