+++ app/src/internal/res/layout/playground_layout.xml
@@ -48,19 +48,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-    <TextView
-        android:id="@+id/text"
-        app:layout_constraintTop_toBottomOf="@id/stepper"
-        android:text="testing"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
-
-    <TextView
-        android:id="@+id/text2"
-        app:layout_constraintTop_toBottomOf="@id/text"
-        android:text="testing"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
   </androidx.constraintlayout.widget.ConstraintLayout>
 
   <Button
